added auth and user interfaces and confirmed registrationform functionality
