+++ src/types/user.types.ts
@@ -1,0 +1,9 @@
+// Output interfaces
+export interface UserProfile {
+    id: number;
+    username: string;
+    email: string;
+    first_name: string;
+    last_name: string;
+    is_staff: boolean;
+    }
